project1/stockanalysisproject/requirements.txt
mysql-connector-python
pandas
plotly
